backend/requirements.txt
fastapi
uvicorn
gphoto2
pillow
pydantic
python-multipart
sh
opencv-python-headless
rawpy
imageio
